+++ examples/supabase/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @stackframe/example-supabase
 
+## 2.8.38
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack@2.8.38
+
+## 2.8.37
+
+### Patch Changes
+
+- @stackframe/stack@2.8.37
+
+## 2.8.36
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.36
+
 ## 2.8.35
 
 ### Patch Changes