+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @stackframe/init-stack
 
+## 2.8.38
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.38
+
+## 2.8.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.37
+
+## 2.8.36
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.36
+
 ## 2.8.35
 
 ### Patch Changes