+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-shared
 
+## 2.8.38
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.37
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.36
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.35
 
 ### Patch Changes